slå sammen identiske DetaljertResultatTidslinjeUtledere til én
Ung- og aktivitetspenger-utlederne var nå identiske. Erstatter dem med én felles
DefaultDetaljertResultatTidslinjeUtleder i formidling, registrert for begge ytelser
via @FagsakYtelseTypeRef. Ytelse-spesifikk ProsessTriggerPeriodeUtleder slås opp per
behandling (samme mønster som ProsessTriggerPeriodeUtleder.finnTjeneste).

- Ny felles DetaljertResultatPeriodeGrunnlag i formidling erstatter
  UngDetaljertResultatPeriodeGrunnlag og AktivitetspengerDetaljertResultatGrunnlag.
- Fjerner UngDetaljertResultatTidslinjeUtleder og
  AktivitetspengerDetaljertResultatTidslinjeUtleder.
