+++ formidling/src/main/java/no/nav/ung/sak/formidling/vedtak/resultat/DetaljertResultatPeriodeGrunnlag.java
@@ -1,0 +1,31 @@
+package no.nav.ung.sak.formidling.vedtak.resultat;
+
+import no.nav.ung.kodeverk.behandling.BehandlingÅrsakType;
+import no.nav.ung.kodeverk.vilkår.Utfall;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+/**
+ * Mellomobjekt for vilkårsresultat og behandlingsårsaker per periode, primært brukt for tidslinje
+ * i {@link DetaljertResultatTidslinjeUtleder}.
+ */
+public record DetaljertResultatPeriodeGrunnlag(List<DetaljertVilkårResultat> vilkårsresultater,
+                                               Set<BehandlingÅrsakType> behandlingÅrsaker,
+                                               boolean tilVurdering) {
+
+    public Set<DetaljertVilkårResultat> avslåtteVilkår() {
+        return vilkårsresultater.stream()
+            .filter(it -> it.utfall() == Utfall.IKKE_OPPFYLT)
+            .collect(Collectors.toSet());
+    }
+
+    public Set<DetaljertVilkårResultat> ikkeVurderteVilkår() {
+        return vilkårsresultater.stream()
+            .filter(it -> it.utfall() == Utfall.IKKE_VURDERT)
+            .collect(Collectors.toSet());
+    }
+
+}
+
