--- ytelse-aktivitetspenger/src/main/java/no/nav/ung/ytelse/aktivitetspenger/formidling/vedtak/resultat/AktivitetspengerDetaljertResultatGrunnlag.java
@@ -1,30 +1,0 @@
-package no.nav.ung.ytelse.aktivitetspenger.formidling.vedtak.resultat;
-
-import no.nav.ung.kodeverk.behandling.BehandlingÅrsakType;
-import no.nav.ung.kodeverk.vilkår.Utfall;
-import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertVilkårResultat;
-
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-/**
- *
- * Intermidiate object for vilkårresultat and behandling årsaker primært brukt for tidslinje
- *
- */
-public record AktivitetspengerDetaljertResultatGrunnlag(List<DetaljertVilkårResultat> vilkårsresultater,
-                                                        Set<BehandlingÅrsakType> behandlingÅrsaker,
-                                                        boolean tilVurdering) {
-
-    public Set<DetaljertVilkårResultat> avslåtteVilkår() {
-        return vilkårsresultater.stream()
-            .filter(it -> it.utfall() == Utfall.IKKE_OPPFYLT)
-            .collect(Collectors.toSet());
-    }
-
-    public Set<DetaljertVilkårResultat> ikkeVurderteVilkår() {
-        return vilkårsresultater.stream().filter(it -> it.utfall() == Utfall.IKKE_VURDERT).collect(Collectors.toSet());
-    }
-
-}
